refactor(auto-review): decompõe AutoReviewFieldPanel (react-doctor no-giant-component)
useAutoReviewKeyboard concentra o listener global de teclado (1-4/P/N) e o
auto-advance pós-decisão (timeout de 250ms), sem mudança de comportamento —
segue remontando com o painel via key={currentKey}. AutoReviewVerdictButtons,
AutoReviewJustificationInput (com o effect de foco) e AutoReviewFooter
(atalhos + status/submit) viram componentes próprios. O override
no-event-handler do doctor.config.json migra de AutoReviewFieldPanel.tsx
para AutoReviewJustificationInput.tsx, que herdou o effect suprimido.

Refs #357
